master page review

+++ IBS/DataAccess/T101IeCluster.cs
@@ -9,7 +9,7 @@ public partial class T101IeCluster
 
     public string DepartmentCode { get; set; } = null!;
 
-    public byte ClusterCode { get; set; }
+    public int ClusterCode { get; set; }
 
     public string? UserId { get; set; }
